minimized the express server
