+++ package.json
@@ -3,26 +3,26 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "start": "webpack && DEBUG=pl_redux:* node ./bin/www"
+    "start": "concurrent \"npm run server\" \"npm run webpack\" ",
+    "server": "node ./server/app.js",
+    "webpack": "webpack -w"
   },
   "dependencies": {
-    "body-parser": "~1.13.2",
-    "cookie-parser": "~1.3.5",
-    "debug": "~2.2.0",
+    "debug": "^2.2.0",
     "ejs": "~2.3.3",
     "express": "~4.13.1",
-    "morgan": "~1.6.1",
-    "serve-favicon": "~2.3.0"
+    "morgan": "^1.7.0"
   },
   "devDependencies": {
     "babel-core": "^6.5.2",
     "babel-loader": "^6.2.4",
     "babel-preset-es2015": "^6.5.0",
     "babel-preset-react": "^6.5.0",
+    "concurrently": "^2.0.0",
     "jquery": "^2.2.1",
     "lodash.debounce": "^4.0.3",
     "react": "^0.14.7",
-    "webpack": "^1.12.14",
-    "webpack-dev-server": "^1.14.1"
+    "react-dom": "^0.14.7",
+    "webpack": "^1.12.14"
   }
 }
