+++ server/views/index.ejs
@@ -1,10 +1,11 @@
 <!DOCTYPE html>
 <html>
   <head>
-    <title><%= title %></title>
+    <title>Currency Converter</title>
     <link rel='stylesheet' href='/css/style.css' />
   </head>
   <body>
+    <h1>Currency Converter</h1>
     <div id="container"></div>
     <script src="/js/build/bundle.js"></script>
   </body>
